Correcciones en Activity

+++ Entity/DTO/PersonDTO.cs
@@ -10,7 +10,7 @@ namespace Entity.DTO
         public string? FirstName { get; set; }
         public string? LastName { get; set; }
         public int Document { get; set; }
-        public int PhoneNumber { get; set; }
+        public long PhoneNumber { get; set; }
         public string? Email { get; set; }
     }
 }
